requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.7.0
base58==2.1.1
bcrypt==3.2.0
cffi==1.15.1
click==8.1.3
dnspython==2.1.0
ecdsa==0.18.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
Flask==2.2.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==2.5.1
passlib==1.7.4
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.3
six==1.16.0
sniffio==1.3.0
starlette==0.16.0
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.14.0
Werkzeug==2.2.2
zipp==3.9.0
